+++ django_project/gap/admin/__init__.py
@@ -5,5 +5,6 @@ Tomorrow Now GAP.
 .. note:: Admins
 """
 
+from gap.admin.crop_insight import *
 from gap.admin.farm import *
 from gap.admin.main import *
