+++ django_project/gap/factories/main.py
@@ -10,7 +10,6 @@ from django.contrib.gis.geos import Point, MultiPolygon, Polygon
 from core.factories import BaseMetaFactory, BaseFactory
 from gap.models import (
     CastType,
-    Crop,
     DatasetType,
     Dataset,
     Provider,
@@ -198,20 +197,6 @@ class DataSourceFileFactory(
     end_date_time = factory.Faker('date_time')
     created_on = factory.Faker('date_time')
     format = DatasetStore.NETCDF
-
-
-class CropFactory(
-    BaseFactory[Crop], metaclass=BaseMetaFactory[Crop]
-):
-    """Factory class for Crop model."""
-
-    class Meta:  # noqa
-        model = Crop
-
-    name = factory.Sequence(
-        lambda n: f'crop-{n}'
-    )
-    description = factory.Faker('text')
 
 
 class VillageFactory(
