+++ django_project/gap/models/__init__.py
@@ -10,4 +10,5 @@ from gap.models.dataset import *
 from gap.models.farm import *
 from gap.models.ingestor import *
 from gap.models.measurement import *
+from gap.models.crop_insight import *
 from gap.models.station import *
