+++ django_project/gap/models/crop_insight.py
@@ -1,0 +1,14 @@
+# coding=utf-8
+"""
+Tomorrow Now GAP.
+
+.. note:: Models
+"""
+
+from core.models.common import Definition
+
+
+class Crop(Definition):
+    """Model representing crop."""
+
+    pass
