+++ django_project/gap/tests/crop_insight/__init__.py
@@ -1,0 +1,6 @@
+# coding=utf-8
+"""
+Tomorrow Now GAP.
+
+.. note:: Plant tests
+"""
