+++ django_project/spw/tests/test_generator.py
@@ -5,33 +5,34 @@ Tomorrow Now GAP.
 .. note:: UnitTest for Plumber functions.
 """
 from datetime import datetime, timedelta
-import pytz
 from unittest.mock import patch, MagicMock
-from django.test import TestCase
+
+import pytz
 from django.contrib.gis.geos import Point
+from django.test import TestCase
 
 from gap.models import (
     DatasetAttribute,
     Dataset
+)
+from gap.providers.tio import (
+    TomorrowIODatasetReader
 )
 from gap.utils.reader import (
     DatasetReaderInput,
     DatasetReaderValue,
     DatasetTimelineValue
 )
-from gap.providers.tio import (
-    TomorrowIODatasetReader
+from spw.factories import (
+    RModelFactory
 )
-from spw.models import RModelExecutionLog, RModelExecutionStatus
 from spw.generator import (
     SPWOutput,
     _fetch_timelines_data,
     _fetch_ltn_data,
     calculate_from_point
 )
-from spw.factories import (
-    RModelFactory
-)
+from spw.models import RModelExecutionLog, RModelExecutionStatus
 
 
 class TestSPWOutput(TestCase):
@@ -143,16 +144,18 @@ class TestSPWFetchDataFunctions(TestCase):
             ])
         )
         result = _fetch_timelines_data(
-            self.location_input, self.attrs, self.start_dt, self.end_dt)
+            self.location_input, self.attrs, self.start_dt, self.end_dt
+        )
         expected_result = {
             '07-20': {
                 'date': '2023-07-20',
                 'evapotranspirationSum': 10,
-                'rainAccumulationSum': 5
+                'rainAccumulationSum': 5,
+                'temperatureMax': 0,
+                'temperatureMin': 0
             }
         }
         self.assertEqual(result, expected_result)
-
 
     @patch.object(TomorrowIODatasetReader, 'read')
     @patch.object(TomorrowIODatasetReader, 'get_data_values')
@@ -202,8 +205,8 @@ class TestSPWGenerator(TestCase):
     @patch('spw.generator._fetch_timelines_data')
     @patch('spw.generator._fetch_ltn_data')
     def test_calculate_from_point(
-        self, mock_fetch_ltn_data, mock_fetch_timelines_data,
-        mock_execute_spw_model):
+            self, mock_fetch_ltn_data, mock_fetch_timelines_data,
+            mock_execute_spw_model):
         """Test calculate_from_point function."""
         mock_fetch_ltn_data.return_value = {
             '07-20': {
